Pawn Refactor (#132)

+++ Code/Player/PlayerController/PlayerPawn.Commands.cs
@@ -1,6 +1,6 @@
 namespace Facepunch;
 
-public partial class PlayerController
+public partial class PlayerPawn
 {
 	/// <summary>
 	/// Development: should bots follow the player's input?
@@ -10,35 +10,40 @@ public partial class PlayerController
 	[DeveloperCommand( "-10 HP (head)", "Player" )]
 	private static void Command_HurtTenHead()
 	{
-		var player = GameUtils.LocalPlayer;
+		var player = PlayerState.Local.PlayerPawn;
+		if ( player is null ) return;
 		player.HealthComponent.TakeDamage( new DamageInfo( player as Component, 10, Hitbox: HitboxTags.Head ) );
 	}
 
 	[DeveloperCommand( "-10 HP (chest)", "Player" )]
 	private static void Command_HurtTenChest()
 	{
-		var player = GameUtils.LocalPlayer;
+		var player = PlayerState.Local.PlayerPawn;
+		if ( player is null ) return;
 		player.HealthComponent.TakeDamage( new DamageInfo( player as Component, 10, Hitbox: HitboxTags.Chest ) );
 	}
 
 	[DeveloperCommand( "Heal", "Player" )]
 	private static void Command_Heal()
 	{
-		var player = GameUtils.LocalPlayer;
+		var player = PlayerState.Local.PlayerPawn;
+		if ( player is null ) return;
 		player.HealthComponent.Health = player.HealthComponent.MaxHealth;
 	}
 
 	[DeveloperCommand( "Suicide", "Player" ), ConCmd( "kill" )]
 	private static void Command_Suicide()
 	{
-		var player = GameUtils.LocalPlayer;
+		var player = PlayerState.Local.PlayerPawn;
+		if ( player is null ) return;
 		player.Kill();
 	}
 
 	[DeveloperCommand( "Give $1k", "Player" )]
 	private static void Command_GiveGrand()
 	{
-		var player = GameUtils.LocalPlayer;
-		player.Inventory.GiveCash(1000);
+		var player = PlayerState.Local.PlayerPawn;
+		if ( player is null ) return;
+		player.PlayerState.GiveCash(1000);
 	}
 }
